requirements.txt
cffi (1.9.1)
cryptography (1.7.1)
idna (2.2)
pip (9.0.1)
pyasn1 (0.1.9)
pycparser (2.17)
pyenchant (1.6.8)
pyOpenSSL (16.2.0)
requests (2.12.4)
setuptools (32.3.1)
six (1.10.0)
wheel (0.30.0a0)
